fix little mistakes, delete ambigiuous dto
