--- blablacar-api/src/main/java/cz/muni/fi/pa165/blablacar/api/dto/ChangeCommentContentDTO.java
@@ -1,4 +1,0 @@
-package cz.muni.fi.pa165.blablacar.api.dto;
-
-public class ChangeCommentContentDTO {
-}
