+++ blablacar-api/src/main/java/cz/muni/fi/pa165/blablacar/api/dto/comment/CommentDTO.java
@@ -1,4 +1,6 @@
-package cz.muni.fi.pa165.blablacar.api.dto;
+package cz.muni.fi.pa165.blablacar.api.dto.comment;
+
+import cz.muni.fi.pa165.blablacar.api.dto.DriveDTO;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
